moved smc_step to smc folder and updated all imports
